Testing updates. Message looping robusting
