+++ unit_test/CMakeLists.txt
@@ -1,18 +1,21 @@
 set(DEVICE_TEST_FILES "tests/Errors.cpp" 
                       "tests/Utilities.cpp"
+                      "tests/Subsystems/Model_Subsystem.cpp" 
+                      "tests/Subsystems/View_Subsystem.cpp" 
+                      "tests/Subsystems/Controller_Subsystem.cpp"
                       "tests/Device.cpp"
                       "tests/Node.cpp"
                       "tests/Network.cpp"
                       "tests/Commands.cpp"
-                      "tests/Subsystem.cpp"
                       "tests/Network_Messages.cpp" 
                       "tests/Internal_Messages.cpp"
-                      "tests/Command_View_Message.cpp" 
-                      "tests/Action_Layer.cpp"
-                      "tests/Action_Stack.cpp"
-                      "tests/Keyboard_Buffer_Input.cpp" 
-                      "tests/Keyboard_Hardware.cpp" 
-                      "tests/Keyboard_System_Not_Connected.cpp")
+                      "tests/View/Command_View_Message.cpp"
+                      "tests/View/Option_Pop_Up.cpp"
+                      "tests/Input/Action_Layer.cpp"
+                      "tests/Input/Action_Stack.cpp"
+                      "tests/Input/Keyboard_Buffer_Input.cpp" 
+                      "tests/Input/Keyboard_Hardware.cpp" 
+                      "tests/Input/Keyboard_System_Not_Connected.cpp" )
 
 add_library(Test_Utilities "test_utilities/pch.h"
                            "test_utilities/system_utilities.h"
